Extract Task interface
